Flexmessage作業ファイルをフォルダへ格納
