+++ template_message.rb
@@ -1,0 +1,15 @@
+module LineBot
+  module Messages
+    class TemplateMessage
+      include Linebot::Messages::Concern::Carouselable
+
+      def send
+        carousel('alter_text', [bubble])
+      end
+
+      def bubble
+
+      end
+    end
+  end
+end
