+++ try_flexmessage/carouselable.rb
